net, phases: Move to pkg/network/setup
