--- pkg/network/BUILD.bazel
@@ -1,8 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["reserved_macs.go"],
-    importpath = "kubevirt.io/kubevirt/pkg/network",
-    visibility = ["//visibility:public"],
-)
