+++ pkg/network/infraconfigurators/BUILD.bazel
@@ -11,7 +11,6 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/network/infraconfigurators",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/network:go_default_library",
         "//pkg/network/cache:go_default_library",
         "//pkg/network/driver:go_default_library",
         "//pkg/network/istio:go_default_library",
