+++ pkg/network/link/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "address.go",
         "ethtool.go",
         "names.go",
+        "reserved_macs.go",
     ],
     importpath = "kubevirt.io/kubevirt/pkg/network/link",
     visibility = ["//visibility:public"],
