+++ pkg/network/setup/BUILD.bazel
@@ -7,7 +7,7 @@ go_library(
         "podinterface.go",
         "podnic.go",
     ],
-    importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/network",
+    importpath = "kubevirt.io/kubevirt/pkg/network/setup",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/network/cache:go_default_library",
@@ -35,13 +35,13 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//pkg/ephemeral-disk-utils:go_default_library",
-        "//pkg/network:go_default_library",
         "//pkg/network/cache:go_default_library",
         "//pkg/network/dhcp:go_default_library",
         "//pkg/network/driver:go_default_library",
         "//pkg/network/errors:go_default_library",
         "//pkg/network/infraconfigurators:go_default_library",
         "//pkg/network/istio:go_default_library",
+        "//pkg/network/link:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
