+++ pkg/virt-launcher/virtwrap/BUILD.bazel
@@ -22,6 +22,7 @@ go_library(
         "//pkg/host-disk:go_default_library",
         "//pkg/ignition:go_default_library",
         "//pkg/network/cache:go_default_library",
+        "//pkg/network/setup:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/net/ip:go_default_library",
         "//pkg/virt-handler/cmd-client:go_default_library",
@@ -36,7 +37,6 @@ go_library(
         "//pkg/virt-launcher/virtwrap/device/sriov:go_default_library",
         "//pkg/virt-launcher/virtwrap/efi:go_default_library",
         "//pkg/virt-launcher/virtwrap/errors:go_default_library",
-        "//pkg/virt-launcher/virtwrap/network:go_default_library",
         "//pkg/virt-launcher/virtwrap/stats:go_default_library",
         "//pkg/virt-launcher/virtwrap/util:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
